+++ packages/utils-nest/src/auth/guards/index.ts
@@ -2,4 +2,3 @@ export * from './admin-guard'
 export * from './cred-decorator'
 export * from './cred-guard'
 export * from './cred-types'
-export * from './token.provider.decorator'
