+++ packages/utils-nest/src/auth/index.ts
@@ -3,3 +3,4 @@ export * from './ownership-provider'
 export * from './guards'
 export * from './helpers'
 export * from './auth-provider-interface'
+export * from './token.provider.decorator'
